+++ cli/src/connect/intrinsics.ts
@@ -3,6 +3,7 @@ import { InternalError, ParserContext, ParserError } from './parser_common'
 import {
   assertIsArrayLiteralExpression,
   assertIsStringLiteral,
+  isUndefinedType,
   stripQuotesFromNode,
 } from '../typescript/compiler'
 import { convertObjectLiteralToJs } from '../typescript/compiler'
@@ -505,7 +506,9 @@ export function intrinsicToString(
       // `const propName = figma.properties.instance('propName')`
       if (modifiers.length > 0) {
         const instance = `${selector}.__properties__.__instance__('${args.figmaPropName}')`
-        return [instance, ...modifiers.map(modifierToString)].join('.')
+        let body = `const instance = ${instance}\n`
+        body += `return instance && instance.type !== "ERROR" ? ${['instance', ...modifiers.map(modifierToString)].join('.')} : instance`
+        return `(function () {${body}})()`
       }
       return `${selector}.__properties__.instance('${args.figmaPropName}')`
     }
@@ -555,6 +558,15 @@ ${Object.entries(args.props).map(
   }
 }
 
+/**
+ * Converts an expression to an FCC value, which is a wrapper around the actual value that
+ * includes the type information. This is used to serialize the value to JSON and then
+ * deserialize it back to the correct type in the generated code.
+ *
+ * @param valueNode
+ * @param parserContext
+ * @returns
+ */
 function expressionToFccEnumValue(
   valueNode: ts.Expression,
   parserContext: ParserContext,
@@ -581,7 +593,14 @@ function expressionToFccEnumValue(
     return _fcc_templateString(str)
   }
 
+  // Handles enums, for example `MyEnum.Value`
   if (ts.isPropertyAccessExpression(valueNode)) {
+    return _fcc_identifier(valueNode.getText())
+  }
+
+  // Any other identifiers (except undefined) are treated as React components, for example `MyComponent`.
+  // We don't support referencing other variables in props object so this should be fine.
+  if (ts.isIdentifier(valueNode) && !isUndefinedType(valueNode, parserContext.checker)) {
     return _fcc_identifier(valueNode.getText())
   }